move extractors to eval_clients/extractor
