+++ src/langcheck/metrics/eval_clients/_openrouter.py
@@ -9,7 +9,8 @@ import requests
 from langcheck.utils.progress_bar import tqdm_wrapper
 
 from ..prompts._utils import get_template
-from ._base import EvalClient, Extractor
+from ._base import EvalClient
+from .extractor import Extractor
 
 
 class OpenRouterEvalClient(EvalClient):
