+++ src/langcheck/metrics/eval_clients/extractor/__init__.py
@@ -1,0 +1,9 @@
+from langcheck.metrics.eval_clients.extractor._base import Extractor
+from langcheck.metrics.eval_clients.extractor.string_match_extractor import (
+    StringMatchExtractor,
+)
+
+__all__ = [
+    "Extractor",
+    "StringMatchExtractor",
+]
